requirements.txt
django 
django-reversion 
markdown 
flup
